feat(ZNTA-1302): wip

+++ server/zanata-war/src/main/java/org/zanata/action/ReindexAction.java
@@ -141,7 +141,7 @@ public class ReindexAction implements Serializable {
         return false;
     }
 
-    public int getReindexCount() {
+    public long getReindexCount() {
         if (searchIndexManager.getProcessHandle() == null) {
             return 0;
         } else {
@@ -149,7 +149,7 @@ public class ReindexAction implements Serializable {
         }
     }
 
-    public int getReindexProgress() {
+    public long getReindexProgress() {
         if (searchIndexManager.getProcessHandle() == null) {
             return 0;
         } else {
